+++ src/controllers/UserController.js
@@ -15,7 +15,7 @@ module.exports = {
 
     async create(req, res) {
         try {
-            const { name, username, password, profile_id} = req.body;
+            const { name, username, password, profile_id } = req.body;
 
             const [userCreated, created] = await userService.findOrCreate({
                 name,
